ccm-forum, incorporating: r1630 | chrisg23 | 2007-09-17 10:28:18 +0200 (Mo, 17 Sep 2007) | 1 line
Added:
   A /trunk/ccm-core/web/assets/fckeditor/config/fckconfig_forum.js

FCK Config for forum application - disables browsing of server file system or aplaws content centre when addign a link as this may be used by public - to use this, need to specify com.arsdigita.forum.use_wysiwyg_editor=true. If HTML Area is set as your editor, then browsing is already disabled for forum in the foorum code. If you have set FCK as your editor then this config will be used.


git-svn-id: https://svn.libreccm.org/ccm/trunk@11 8810af33-2d31-482b-a856-94f89814c4df