incorporating APLAWS patch: r1663 | chrisg23 | 2007-09-19 10:23:54 +0200 (Mi, 19 Sep 2007)
Add association between image and link. Changes have no effect by themselves, but if the ccm-wsx-image-link-step module in contrib is loaded then the default image step is replaced with an asset step UI for associating the article image with one of the related links defined for the article. This change requires upgrade script to be run: ccm upgrade ccm-ldn-image-step --from-version 6.5.0 --to-version 6.5.1


git-svn-id: https://svn.libreccm.org/ccm/trunk@27 8810af33-2d31-482b-a856-94f89814c4df