incorporating APLAWS patches 166-168:
r1666 | chrisg23 | 2007-09-19 12:40:53 +0200 (Mi, 19 Sep 2007) 
Currently date order categories work fine unless you have more than one page of objects - as the ordering is only done in xsl, when you have pagination, the java side returns a subset of objects ordered according to order set in jsp page (eg alphabetical) and then that subset is date ordered, but that means that as you page back and forth the date ordering is completely wrong. This change fixes that by enforcing date ordering in the java code if the current category is date ordered. Different object types have different attribute names for date - and only the renderer knows which attributes are available. I have added a method to the definition that gives the definition a reference to the renderer - the renderer now returns a list of attributes which the definition looks through and finds the first attribute that is of type date and uses that as the basis for date ordering if the current category is date ordered.If there is more than one date ordered attribute output in the list, then ensure the one used for ordering in xsl is the first one specified in jsp. I could have just made a setDateOrderAttribute method that takes a string attribute name argument and called that from the jsp page - it would have made the code quite a bit simpler. I chose this route to ensure that the date attribute specified for ordering in java was present in the datacollection, and so that there is no need to edit existing jsp pages. That is probably being overcautious though. Never mind, this works fine anyway - maybe one day I will refactor it to use the simpler approach
------------------------------------------------------------------------
r1667 | chrisg23 | 2007-09-19 12:48:36 +0200 (Mi, 19 Sep 2007) 
Continuation of changes in r1651 - output publisher flag on index pages too
------------------------------------------------------------------------
r1668 | chrisg23 | 2007-09-19 12:53:08 +0200 (Mi, 19 Sep 2007) 
See r1661 - ensure that this version of object-list.xsl caters for all listed objects, not just content pages, by referring to displayName (defined in ACSObject and now output on all object lists in default navigation jsp page


git-svn-id: https://svn.libreccm.org/ccm/trunk@28 8810af33-2d31-482b-a856-94f89814c4df