incporporating APLAWS patches 1686 & 1687:
r1686 | chrisg23 | 2007-10-23 16:11:16 +0200 (Di, 23 Okt 2007) 
Change sql to a version that both Oracle and Postgres understand
------------------------------------------------------------------------
r1687 | chrisg23 | 2007-10-23 16:16:50 +0200 (Di, 23 Okt 2007) 
sourceforge patch [1803373] - don't stream out blob data until it's actually needed. Note this is configurable so that by default there is no behaviour change, but really it would be unusual to actively want the overhead of the additional streamout. Credit to matt at magpie for this one


git-svn-id: https://svn.libreccm.org/ccm/trunk@35 8810af33-2d31-482b-a856-94f89814c4df