Integrate vendor release   r1795: First part of the conversion of CCM to be deployed in a standard way   r1796: Multilanguage categorization (already local, but upgrade script          added, comments and translation added)
First part of the conversion of CCM / APLAWS to be deployed in a standard way 
by a *.war file.

Backwards compatible to the procedures so far. The ccm scripts can be used
as used to. Only difference is the location of some files (see below).

As an alternative the webapps directory created by the ccm scripts can be
moved into a standard tomcat container or can be added as a virtual host.
APLAWS no longer depends on all the environment variables. For detailed 
instructions see documentation on the WIKI. This is a temporary solution,
further enhancements are on the way!

No update script required for this part.

Modifications:

-   Creation of a c.ad.runtime.CCMResourceManager class, which currently
handles just the file locations. It replaces the dependency from the
CCM_HOME environment variable. It is initialized during startup of the
servlet container or by the entry point of the installation tools (i.e.
c.ad.packaging.mastertool). It replaces the c.ad.packaging.CCM class
as well. Currently contains backwards compatibility code to check the 
runtime environment just in case the initialization is missing (will be 
removed later).

-   Creation of c.ad.web.CCMApplicationContextListener class which is
notified by the servlet container at the beginning of the startup procedure.
Initialises CCMResourceManager. Must be configured by web.xml (if a contributed
package overwrites the trunk values)!

-   File location changes:
    -   Configuration registry moved to [appbase]/WEB-INF/conf/registry
    -   ps2fs files moved to [appbase]/ps2fs
        (can be modified by configuration parameter)
    -   lucene index files moved to [appbase]/WEB-INF/work/lucene
        (can be modified by configuration parameter)
Developers watch out! ant clean now removes these files, too (esp. registry)!

-   XML factory classes are now determined by configuration parameters
    - waf.xml.xsl_transformer=[saxon (default)|jd.xslt|resin|xalan|xsltc]
    - waf.xml.dom_builder=[xerces (default) | resin]
    - waf.xml.sax_parser=[xerces (default) | resin] 

-   Package c.ad.packaging reorganized so it contains just installation
related code. Runtime configuration related code moved to c.ad.runtime.
Code is not part of the API.

-   log4j configuration file is still at the old location, but is not used
anymore! It will be moved to [appbase]/WEB-INF/conf/ in the future.
Developers, modify the file ~/webapps/WEB-INF//classes/log4j.properties 
in the meantime.

-   While working on the code some deprecated classes and method invocations 
replaced by supported API. If a class is no longer used by any module in
trunk, the class has been removed by appending "nolongerInUse" to the file 
name. If any of the contributed modules is still depending on deprecated code
it is easy to find and to follow the recommended replacement instructions. Do
not re-incorporate deprecated code! We have to clean up the code base!


Second part: Multilanguage categories
Added update script
Translation of some comments, added comments
version number bumped to reflect the major changes




git-svn-id: https://svn.libreccm.org/ccm/trunk@73 8810af33-2d31-482b-a856-94f89814c4df